SOX-66 Refactor setup to method to allow reusing - extract deleteBranchesByPrefix
